Authorize application staff comments using gate policy [sc-41]

+++ app/Http/Livewire/Recruitment/ShowApplication.php
@@ -88,7 +88,11 @@ class ShowApplication extends Component
     {
         $comment = Comment::where('uuid', $uuid)->firstOrFail();
 
+        $this->authorize('delete', $comment);
+
         $comment->delete();
+
+        $this->sendDiscordWebhook('Application Comment Deleted', 'By **' . Auth::user()->username . '**', 14429954);
 
         session()->now('alert', ['type' => 'info', 'message' => 'Comment deleted!']);
     }
